review controller fixes

+++ backend/src/main/kotlin/com/jetbrains/life_science/article/review/request/service/ReviewRequestServiceImpl.kt
@@ -31,6 +31,9 @@ class ReviewRequestServiceImpl(
     }
 
     override fun delete(request: ReviewRequest) {
+        if (request.resolution != null) {
+            throw ReviewResponseAlreadyExistsException("This review request already proceed")
+        }
         repository.delete(request)
     }
